hntitles/requirements.txt
blurhash==1.1.4
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.1
et-xmlfile==1.1.0
idna==3.4
Mastodon.py==1.7.0
numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221124
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.6
requests==2.28.1
six==1.16.0
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
